% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosineSimilarity.R
\name{cosine_similarity_matrix_rcpp}
\alias{cosine_similarity_matrix_rcpp}
\title{Compute Cosine Similarity Matrix (Rcpp)}
\usage{
cosine_similarity_matrix_rcpp(mat)
}
\arguments{
\item{mat}{A numeric matrix where rows represent observations}
}
\value{
A symmetric matrix of cosine similarities, with diagnal elements be 1
}
\description{
Calculates pairwise cosine similarity matrix for rows of a matrix using Rcpp.
}
\examples{
mat <- matrix(rnorm(12), nrow = 4, ncol = 3)
cosine_similarity_matrix_rcpp(mat)
}
