% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosineSimilarity.R
\name{cosine_similarity_rcpp}
\alias{cosine_similarity_rcpp}
\title{Compute Cosine Similarity between Two Vectors (Rcpp)}
\usage{
cosine_similarity_rcpp(x, y)
}
\arguments{
\item{x, y}{Numeric vectors of the same length}
}
\value{
Cosine similarity value x and y, value between -1 and 1
}
\description{
Calculates the cosine similarity between two numeric vectors using Rcpp.
}
\examples{
x <- c(1, 2, 3)
y <- c(4, 5, 6)
cosine_similarity_rcpp(x, y)
}
